requirements.txt
pydantic==2.12.5
python-dotenv==1.2.2
httpx==0.28.1
openai==2.29.0
tavily-python==0.7.23
beautifulsoup4==4.14.3
fastapi==0.135.1
uvicorn==0.42.0
pytest==9.0.2